+++ src/test/java/com/epam/tc/hw4/ex1/tests/HomePageTest.java
@@ -3,6 +3,7 @@ package com.epam.tc.hw4.ex1.tests;
 import static org.testng.AssertJUnit.assertEquals;
 import static org.testng.AssertJUnit.assertTrue;
 
+import com.epam.tc.hw4.LoginPage;
 import com.epam.tc.hw4.TestBase;
 import com.epam.tc.hw4.ex1.pages.HomePage;
 import io.qameta.allure.Feature;
@@ -15,11 +16,13 @@ import org.testng.annotations.Test;
 
 public class HomePageTest extends TestBase {
     HomePage homePage;
+    LoginPage loginPage;
 
     @BeforeMethod
     public void setup() {
         init();
         homePage = new HomePage();
+        LoginPage loginPage;
     }
 
     @Feature(value = "Homework 4")
@@ -53,14 +56,13 @@ public class HomePageTest extends TestBase {
 
     @Step("Perform login")
     public void userLogin() {
-        homePage.clickOnUserIcon();
-        homePage.login(prop.getProperty("username"), prop.getProperty("password"));
+        loginPage.login(prop.getProperty("username"), prop.getProperty("password"));
     }
 
     @Step("Assert Username is loggined")
     public void checkUserIsLogined() {
-        assertTrue(homePage.isUserNameDisplayed());
-        assertTrue(homePage.getUserName().getText()
+        assertTrue(loginPage.isUserNameDisplayed());
+        assertTrue(loginPage.getUserName().getText()
                 .contains(prop.getProperty("fullusername")));
     }
