--- src/test/resources/HomePageTestSuite.xml
@@ -1,8 +1,0 @@
-<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd" >
-<suite name="Home Page Test Suite">
-    <test name="Home Page Test">
-        <classes>
-            <class name="com.epam.tc.hw4.ex1.tests.HomePageTest"/>
-        </classes>
-    </test>
-</suite>
