--- src/test/resources/ServicePageTestSuite.xml
@@ -1,8 +1,0 @@
-<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd" >
-<suite name="Service Page Test Suite">
-    <test name="Service Page Test">
-        <classes>
-            <class name="com.epam.tc.hw4.ex2.tests.ServicePageTest"/>
-        </classes>
-    </test>
-</suite>
